created Training Resource Dao
