+++ Second-Nature/java/src/main/java/com/techelevator/model/TrainingResource/ArchiveTrainingResourceDto.java
@@ -1,4 +1,4 @@
-package com.techelevator.model;
+package com.techelevator.model.TrainingResource;
 
 public class ArchiveTrainingResourceDto {
     private int trainingResourceId;
