+++ Second-Nature/java/src/main/java/com/techelevator/model/TrainingResource/TrainingResource.java
@@ -1,4 +1,4 @@
-package com.techelevator.model;
+package com.techelevator.model.TrainingResource;
 
 import java.sql.Timestamp;
 import java.util.Date;
