+++ Second-Nature/java/src/main/java/com/techelevator/model/TrainingResource/UpdateTrainingResourceDto.java
@@ -1,9 +1,18 @@
-package com.techelevator.model;
+package com.techelevator.model.TrainingResource;
 
-public class CreateTrainingResourceDTO {
+public class UpdateTrainingResourceDto {
+    private int trainingResourceId;
     private String category;
     private String content;
     private String resourceSource;
+
+    public int getTrainingResourceId() {
+        return trainingResourceId;
+    }
+
+    public void setTrainingResourceId(int trainingResourceId) {
+        this.trainingResourceId = trainingResourceId;
+    }
 
     public String getCategory() {
         return category;
